Move filter methods to each class
