+++ src/graphql/queries/search-events.js
@@ -1,8 +1,54 @@
-const {
-  buildCursorOptions,
-  buildEventFilters,
-  buildEventSearchPhrase,
-} = require('./utils');
+const { buildCursorOptions } = require('./utils');
+
+const buildFilters = ({
+  OR = [],
+  txid,
+  address,
+  ownerAddress,
+  resultIndex,
+  hashId,
+  status,
+  language,
+}) => {
+  const filter = (
+    txid
+    || address
+    || ownerAddress
+    || resultIndex
+    || hashId
+    || status
+    || language
+  ) ? {} : null;
+
+  if (txid) filter.txid = txid;
+  if (address) filter.address = address;
+  if (ownerAddress) filter.ownerAddress = ownerAddress;
+  if (status) filter.status = status;
+  if (resultIndex) filter.resultIndex = resultIndex;
+  if (hashId) filter.hashId = hashId;
+  if (language) filter.language = language;
+
+  let filters = filter ? [filter] : [];
+  for (let i = 0; i < OR.length; i++) {
+    filters = filters.concat(buildFilters(OR[i]));
+  }
+  return filters;
+};
+
+const buildSearchPhrase = (searchPhrase) => {
+  const filterFields = ['_id', 'name', 'address', 'centralizedOracle'];
+  if (!searchPhrase) return [];
+
+  const filters = [];
+  const searchRegex = new RegExp(`.*${searchPhrase}.*`, 'i');
+  for (let i = 0; i < filterFields.length; i++) {
+    const filter = {};
+    filter[filterFields[i]] = { $regex: searchRegex };
+    filters.push(filter);
+  }
+
+  return filters;
+};
 
 module.exports = async (
   root,
@@ -10,8 +56,8 @@ module.exports = async (
   { db: { Events } },
 ) => {
   const filters = [];
-  if (filter) filters.push({ $or: buildEventFilters(filter) });
-  if (searchPhrase) filters.push({ $or: buildEventSearchPhrase(searchPhrase) });
+  if (filter) filters.push({ $or: buildFilters(filter) });
+  if (searchPhrase) filters.push({ $or: buildSearchPhrase(searchPhrase) });
 
   const query = { $and: filters };
   let cursor = Events.cfind(query);
