+++ src/graphql/queries/utils.js
@@ -20,137 +20,6 @@ const buildCursorOptions = (cursor, orderBy, limit, skip) => {
   return cursor;
 };
 
-const buildEventFilters = ({
-  OR = [],
-  txid,
-  address,
-  ownerAddress,
-  resultIndex,
-  hashId,
-  status,
-  language,
-}) => {
-  const filter = (
-    txid
-    || address
-    || ownerAddress
-    || resultIndex
-    || hashId
-    || status
-    || language
-  ) ? {} : null;
-
-  if (txid) filter.txid = txid;
-  if (address) filter.address = address;
-  if (ownerAddress) filter.ownerAddress = ownerAddress;
-  if (status) filter.status = status;
-  if (resultIndex) filter.resultIndex = resultIndex;
-  if (hashId) filter.hashId = hashId;
-  if (language) filter.language = language;
-
-  let filters = filter ? [filter] : [];
-  for (let i = 0; i < OR.length; i++) {
-    filters = filters.concat(buildEventFilters(OR[i]));
-  }
-  return filters;
-};
-
-const buildEventSearchPhrase = (searchPhrase) => {
-  const filterFields = ['_id', 'name', 'address', 'centralizedOracle'];
-  if (!searchPhrase) return [];
-
-  const filters = [];
-  const searchRegex = new RegExp(`.*${searchPhrase}.*`, 'i');
-  for (let i = 0; i < filterFields.length; i++) {
-    const filter = {};
-    filter[filterFields[i]] = { $regex: searchRegex };
-    filters.push(filter);
-  }
-
-  return filters;
-};
-
-const buildBetFilters = ({
-  OR = [],
-  txid,
-  eventAddress,
-  betterAddress,
-  resultIndex,
-  eventRound,
-}) => {
-  const filter = (
-    txid
-    || eventAddress
-    || betterAddress
-    || resultIndex
-    || eventRound
-  ) ? {} : null;
-
-  if (txid) filter.txid = txid;
-  if (eventAddress) filter.eventAddress = eventAddress;
-  if (betterAddress) filter.betterAddress = betterAddress;
-  if (resultIndex) filter.resultIndex = resultIndex;
-  if (eventRound) filter.eventRound = eventRound;
-
-  let filters = filter ? [filter] : [];
-  for (let i = 0; i < OR.length; i++) {
-    filters = filters.concat(buildBetFilters(OR[i]));
-  }
-  return filters;
-};
-
-const buildResultSetFilters = ({
-  OR = [],
-  txid,
-  eventAddress,
-  centralizedOracleAddress,
-  resultIndex,
-  eventRound,
-}) => {
-  const filter = (
-    txid
-    || eventAddress
-    || centralizedOracleAddress
-    || resultIndex
-    || eventRound
-  ) ? {} : null;
-
-  if (txid) filter.txid = txid;
-  if (eventAddress) filter.eventAddress = eventAddress;
-  if (centralizedOracleAddress) filter.centralizedOracleAddress = centralizedOracleAddress;
-  if (resultIndex) filter.resultIndex = resultIndex;
-  if (eventRound) filter.eventRound = eventRound;
-
-  let filters = filter ? [filter] : [];
-  for (let i = 0; i < OR.length; i++) {
-    filters = filters.concat(buildResultSetFilters(OR[i]));
-  }
-  return filters;
-};
-
-const buildWithdrawFilters = ({
-  OR = [],
-  txid,
-  eventAddress,
-  winnerAddress,
-}) => {
-  const filter = (
-    txid
-    || eventAddress
-    || winnerAddress
-  ) ? {} : null;
-
-  if (txid) filter.txid = txid;
-  if (eventAddress) filter.eventAddress = eventAddress;
-  if (winnerAddress) filter.winnerAddress = winnerAddress;
-
-  let filters = filter ? [filter] : [];
-  for (let i = 0; i < OR.length; i++) {
-    filters = filters.concat(buildWithdrawFilters(OR[i]));
-  }
-  return filters;
-};
-
 const calculateSyncPercent = async (blockNum) => {
   const chainBlock = await web3.eth.getBlock('latest');
   Math.floor((blockNum / chainBlock) * 100);
@@ -185,11 +54,6 @@ const runPaginatedQuery = async ({ db, filter, orderBy, limit, skip }) => {
 
 module.exports = {
   buildCursorOptions,
-  buildEventFilters,
-  buildEventSearchPhrase,
-  buildBetFilters,
-  buildResultSetFilters,
-  buildWithdrawFilters,
   calculateSyncPercent,
   runPaginatedQuery,
 };
